♻️ refactor: 브랜드 스냅 코드 수정, 영어로 변환 코드 분리  #18
- DB에서 다운 완료 된 number 가져와, 크롤링 준비하기
- 파일명 영어로 저장하는 코드 분리
