+++ Flask_backend/crawling/musinsa_codishop_crawling.py
@@ -1,6 +1,7 @@
 import requests
 from bs4 import BeautifulSoup
 import os, sys
+from change_to_english import change_season_eng, change_style_eng
 
 sys.path.append(os.getcwd())
 from DB.DB_setting import connect_to_database
@@ -25,38 +26,7 @@ SCOPES = ["https://www.googleapis.com/auth/drive"]
 save_folderPath = "newimages/20240502/"
 
 # ------------------------------------------ #
-#        계절감과 스타일을 영어로 변환         #
 
-def change_style_eng(style):
-    style_map = {
-        "캐주얼": "casual", 
-        "아메카지": "amekaji", 
-        "시크": "chic", 
-        "스포티": "sporty", 
-        "스트릿": "street", 
-        "비즈니스캐주얼": "businessCasual", 
-        "로맨틱": "romantic", 
-        "레트로": "retro", 
-        "골프": "golf", 
-        "고프코어": "gofcore", 
-        "걸리시": "girlish", 
-        "미니멀": "minimal", 
-        "댄디": "dandy"
-    }
-
-    return style_map.get(style)
-    
-def change_season_eng(season):
-    season_map = {
-        "봄": "spring", 
-        "여름": "summer", 
-        "가을": "autumn", 
-        "겨울": "winter"
-    }
-    
-    return season_map.get(season)
-
-# -------------------------------------------- #
 
 def download_images(url,index):
     gender = "etc"
